+++ const.go
@@ -1,0 +1,5 @@
+package iovns
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+var ZeroAddress = sdk.AccAddress{0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0}