+++ CHANGELOG.md
@@ -4,9 +4,19 @@
 
 - allow empty account name on msg.Validate()
 
-Breaking changes
+## 0.2.3
 
-- change naming of some json keys in genesis.json
+- fix add signers in msg renew account and renew domain
+- add resolve certificates
+- add resolve blockchain targets
+- add generalized indexing strategy
+- abstract indexing 
+- iovnscli: accept certificate as file
+
+### Breaking changes
+
+- Change naming of some json keys in genesis.json
+- change move blockchain address from iovns types to domain module types
 
 ## 0.2.2